Adds step to perform database migration.
